rearrange and fixed most of the errors

+++ ml/indicators.py
@@ -245,16 +245,16 @@ def calculate_obv(data: pd.DataFrame) -> pd.Series:
         logger.warning("Insufficient data for OBV calculation: need at least 2 data points")
         return pd.Series([np.nan] * len(data), index=data.index)
 
-    obv = pd.Series([0] * len(data), index=data.index)
-    obv.iloc[0] = data['volume'].iloc[0]
+    obv = pd.Series([0] * len(data), index=data.index, dtype=int)
+    obv.iloc[0] = int(data['volume'].iloc[0])
 
     for i in range(1, len(data)):
         if data['close'].iloc[i] > data['close'].iloc[i-1]:
-            obv.iloc[i] = obv.iloc[i-1] + data['volume'].iloc[i]
+            obv.iloc[i] = int(obv.iloc[i-1] + data['volume'].iloc[i])
         elif data['close'].iloc[i] < data['close'].iloc[i-1]:
-            obv.iloc[i] = obv.iloc[i-1] - data['volume'].iloc[i]
+            obv.iloc[i] = int(obv.iloc[i-1] - data['volume'].iloc[i])
         else:
-            obv.iloc[i] = obv.iloc[i-1]
+            obv.iloc[i] = int(obv.iloc[i-1])
 
     return obv
